refactor(server): move /api/config routes into config-routes.ts
Keep grid/preset server code out of index.ts (matching main's mount*Routes
pattern) so it doesn't churn the shared file and conflict with unrelated server
changes. index.ts now just calls mountConfigRoutes(app, CLAUDE_CWD).
